refactor(deepseek_v4): factor HyperConnection into shared module + mlx#3448 TODO
Move HyperConnection and HyperHead from deepseek_v4.py into
mlx_lm/models/hyper_connection.py per pcuenca's review on PR #1189.
Matches the shared-module pattern of ssm.py and gated_delta.py — future
mHC adopters can import the primitives directly rather than copy-paste
from the model file. Sinkhorn module already factored separately
(see f49af91), so the extraction is a clean import-rewrite.

Add a TODO(mlx#3448) marker above _reinterpret_safetensor_e8m0_scales_as_uint8
in utils.py — once that upstream PR lands and MLX's safetensors loader
handles F8_E8M0 natively, the byte-reinterpret workaround can be dropped.

No functional changes. 9/9 V4 tests pass.
